+++ packages/core/provider/src/types.ts
@@ -1,0 +1,32 @@
+export type Components = 'button'
+
+export type CommonConfig = {
+  transitionDuration: number | undefined
+}
+
+export type GlobalConfig = {
+  /** Prefix classname for all components.
+   *
+   * Ex: `nsui-button` */
+  prefixCls?: string
+  /** Duration in milliseconds for transitions and animations of all components.
+   *
+   * Defaults to `300ms`. */
+  transitionDuration?: number
+}
+
+export type GlobalConfigRequired = Required<GlobalConfig>
+
+export interface ButtonConfig extends CommonConfig {
+  /** Defines if ripple effect should be disabled
+   *
+   * Defaults to `false`.
+   */
+  isRippleDisabled: boolean
+}
+
+export type NSUIComponentsConfig = {
+  button?: ButtonConfig
+}
+
+export type NSUIComponentsConfigRequired = Required<NSUIComponentsConfig>